Added abstract operator base class. Added new branch-free operator class.
